Moved Contact.js and Contact.css to components folder and changed the coding for the input fields in the contact form
