+++ src/components/Contact.js
@@ -2,22 +2,30 @@ import React, { Component } from "react";
 import "./Contact.css";
 
 class Contact extends Component {
-  state = {
-    firstName: "",
-    lastName: "",
-    phone: "",
-    email: "",
-    message: "",
-  };
+  constructor(props) {
+    super(props);
+    this.state = {
+      firstName: "",
+      lastName: "",
+      phone: "",
+      email: "",
+      message: "",
+    };
+    // To ensure this is calling this.handleChange
+    this.handleChange = this.handleChange.bind(this);
+  }
 
-  handleChange = (e) => {
+  //Field could be firstname, lastname,phone.email or message
+  //value is the content inputted by the user
+  handleChange(field, value) {
     this.setState({
-      [e.target.id]: [e.target.value],
-    });
-  };
+      [field]:value,
+    })
+  }
 
+//On Submitting the form
   handleSubmit = (e) => {
-    // e.preventDefault();
+    e.preventDefault();
     console.log(this.state);
     this.setState({
       firstName: "",
@@ -28,6 +36,7 @@ class Contact extends Component {
     });
   };
 
+  
   render() {
     return (
       <div className="container_contact">
@@ -40,36 +49,48 @@ class Contact extends Component {
           <form onSubmit={this.handleSubmit}>
             <div className="row input-field s6">
               <label htmlFor="firstname">First Name</label>
-              <input type="text" id="firstname" onChange={this.handleChange} />
+              <input
+                type="text"
+                value={this.state.firstName}
+                onChange={(e) => this.handleChange('firstName',e.target.value)}
+              />
             </div>
 
             <div className="row input-field s6">
               <label htmlFor="lastname">Last Name</label>
-              <input type="text" id="lastname" onChange={this.handleChange} />
+              <input
+                type="text"
+                value={this.state.lastName}
+                onChange={(e) => this.handleChange('lastName',e.target.value)}
+              />
             </div>
 
             <div className="row input-field s6">
               <label htmlFor="phone">Phone no:</label>
-              <input type="text" id="phone" onChange={this.handleChange} />
+              <input
+                type="text"
+                value={this.state.phone}
+                onChange={(e) => this.handleChange('phone',e.target.value)}
+              />
             </div>
 
             <div className="row input-field s6">
               <label htmlFor="email">Email:</label>
               <input
                 type="email"
-                id="email"
+                value={this.state.email}
                 className="validate"
-                onChange={this.handleChange}
+                onChange={(e) => this.handleChange('email',e.target.value)}
               />
             </div>
 
             <div className="row input-field s6">
               <label htmlFor="message">Message(optional):</label>
               <textarea
-                name="message"
-                id="message"
+                type="text"
+                value={this.state.message}
                 className="materialize-textarea"
-                onChange={this.handleChange}
+                onChange={(e) => this.handleChange('message',e.target.value)}
               ></textarea>
             </div>
